+++ src/main/java/data_access/IStockDataAccess.java
@@ -1,23 +1,16 @@
 package data_access;
 
+import entity.Stock;
+
 import java.util.Map;
 
 public interface IStockDataAccess {
 
     /**
-     * Get the price of a stock
-     *
-     * @param ticker: the stock ticker of the stock
-     * @return the price of the stock with the given ticker
-     */
-    double getStockPrice(String ticker);
-
-    /**
      * Get the prices of all stocks
      *
-     * @return a hashmap with the stock ticker as the key and the stock price as the value
-     * It should contain all stocks in the database
+     * @return a hashmap with the stock ticker as the key and the Stock entity as the value.
+     * It should contain all stocks in the database.
      */
-    // TODO: get stocks from a list of tickers
-    Map<String, Double> getStocks();
+    Map<String, Stock> getStocks();
 }